refactor: Move ContextGate logic into ApiContext (#19)
